requirements.txt
djangorestframework==3.14.0
Faker==20.1.0
pytest==7.4.3
pytest-django==4.7.0
asgiref==3.7.2
Brotli==1.1.0
colorama==0.4.6
crispy-bootstrap5==2023.10
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-crispy-forms==2.1
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
whitenoise==6.6.0